requirements.txt
# ============================================
# Agentic AI Hands-On Course — All 18 Days
# Dr. Kanthi Kiran Sirra | Sr. AI Engineer | March-April 2026
# ============================================
# Install: pip install -r requirements.txt
# ============================================

# --- Core LLM & Agent Frameworks ---
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
smolagents[litellm]>=1.0.0
crewai>=0.80.0
crewai-tools>=0.14.0

# --- Vector Databases & Embeddings ---
chromadb>=0.5.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# --- RAG & Document Processing ---
pypdf>=4.0.0
unstructured>=0.15.0
rank-bm25>=0.2.2

# --- Evaluation ---
ragas>=0.2.0

# --- Deployment & UI ---
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.30.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.32.0
ddgs>=6.0.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0
pydantic>=2.0.0

# --- Jupyter (for local notebooks if needed) ---
jupyter>=1.0.0
ipykernel>=6.29.0




llama-3.3-70b-versatile